SSHGuard: Implement custom handler for firewall backends
This is the patch to implement custom handler backend for the
SSHGuard tool. This can be useful if you would like to log
all the blocks/releases/flushes into the database using a
separate binary (not included). The only requirement is the
format of the custom handler input arguments.

For reference please see examples/sshguard-custom-script-example.

To be able to use new backends you need to run autoreconf to
generate a new configure script, therefore the configure
script has been deleted. Running 'autoreconf' as the first step
is now required.

To use the custom handler backend you need to provide the custom handler
script/binary location to be used, e.g. like:

./configure --with-firewall=custom --with-custom=/bin/sshguard-custom-script

Hope this will be useful for somebody else, too, and not just me
for me ;-)

Michal

Signed-off-by: Michal Novotny <mignov@gmail.com>